general cleanup before release

+++ src/core/shortcut.cpp
@@ -46,7 +46,8 @@ bool Shortcut::nativeEventFilter(const QByteArray &eventType, void *message, lon
     Q_UNUSED(result);
 
     MSG *msg = static_cast<MSG *>(message);
-    if (msg->message == WM_HOTKEY && msg->wParam == m_id)
+    if (msg->message == WM_HOTKEY
+            && static_cast<int>(msg->wParam) == m_id)
         emit pressed();
 
     return false;
